+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # ??/??/2019
 + Add bootstrap layout
 + Add Public IP on dashboard
++ Enable change password
 
 # 0.1.0
 ## 24-03-2019